chore: rename coordinates and packages
