+++ sample/build.gradle
@@ -14,7 +14,7 @@ repositories {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'com.cloudant:cloudant-spring-boot-starter:0.0.5-SNAPSHOT'
+    implementation 'com.ibm.cloud:cloudant-spring-boot-starter:0.0.5-SNAPSHOT'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
