+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.cloudant.spring.boot.CloudantAutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.cloud.cloudant.spring.boot.CloudantAutoConfiguration
